Redesign video showcase tests for clear narrative flow
Previous video tests had several issues:
- Video config excluded showcase tests (testIgnore matched video dir)
- Tests were fragmented across 6 files with no story arc
- Registration used fast helpers invisible to viewer
- Order placement had ugly polling loop on video

Redesigned as 4 cohesive showcase tests with clear narrative:
1. Home & Guest Browsing — landing page, restaurants, menu, guest prompt
2. Customer Journey — register, browse, order, delivery address, checkout
3. Role Dashboards — restaurant owner + courier registration and dashboards
4. Admin Panel — admin login, dashboard cards, all 4 data tables

Key improvements:
- Video config now targets tests/video/ directly (no testIgnore conflict)
- Forms filled field-by-field with visible pauses for viewer
- Scrolling to reveal full page content at key moments
- Each test ends with logout for clean separation
- No saga polling loop (order submission shown, not backend verification)
